Prefinal improvements

+++ src/routers/review_routers.py
@@ -12,7 +12,8 @@ router_reviews = APIRouter()
 
 
 @router_reviews.post("/reviews", tags=["Reviews"], summary="Создание нового отзыва")
-async def post_new_review(rev_data: ReviewAddDTO):
+async def post_new_review(rev_data: ReviewAddDTO, authorization: str = Header(None)):
+    rev_data.author_id = await Auth.get_user_id(authorization)
     await ReviewORM.insert_review(rev_data)
     return {"ok": True, "message": "Review added successfully"}
 
@@ -45,18 +46,21 @@ async def get_university_reviews(uni_id: int):
     return rev_data
 
 
-@router_reviews.delete("/reviews",
-    tags=["Reviews"], 
-    summary="Удаление отзыва"
-)
-async def delete_review(review_id: int):
-    await ReviewORM.delete_review(review_id)
-    return {"ok": True, "message": "Review deleted successfully"}
+@router_reviews.delete("/reviews", tags=["Reviews"], summary="Удаление отзыва")
+async def delete_review(review_id: int, authorization: str = Header(None)):
+    user_id = await Auth.get_user_id(authorization)
+    if ReviewORM.review_belonging(review_id, user_id):
+        await ReviewORM.delete_review(review_id)
+        return {"ok": True, "message": "Review deleted successfully"}
+    else:
+        raise HTTPException(status_code=403, detail="Forbidden")
 
-@router_reviews.put("/reviews",
-    tags=["Reviews"], 
-    summary="Изменение отзыва"
-)
-async def update_review(review_id: int, review_data: ReviewAddDTO):
-    await ReviewORM.change_review(review_id, review_data)
-    return {"ok": True, "message": "Review updated successfully"}
+
+@router_reviews.put("/reviews", tags=["Reviews"], summary="Изменение отзыва")
+async def update_review(review_id: int, review_data: ReviewAddDTO, authorization: str = Header(None)):
+    user_id = await Auth.get_user_id(authorization)
+    if ReviewORM.review_belonging(review_id, user_id):
+        await ReviewORM.change_review(review_id, review_data)
+        return {"ok": True, "message": "Review updated successfully"}
+    else:
+        raise HTTPException(status_code=403, detail="Forbidden")
